refactor realityMapper
