+++ src/main/java/com/example/demo/inputs/RealityInputAdapter.java
@@ -2,8 +2,8 @@ package com.example.demo.inputs;
 
 import com.example.demo.domain.ports.CreateRealitiesInputPort;
 import com.example.demo.domain.ports.UpdateRealitiesInputPort;
+import com.example.demo.inputs.mappers.RealityInputMapper;
 import com.example.demo.inputs.models.RealityResponse;
-import com.example.demo.reality.RealityMapper;
 import com.example.demo.domain.exceptions.RealityNotFoundException;
 import com.example.demo.domain.models.Reality;
 import com.example.demo.domain.ports.RealitiesInputPort;
@@ -18,7 +18,7 @@ import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @RestController
-@RequestMapping(path="api/v1/realities")
+@RequestMapping(path = "api/v1/realities")
 public class RealityInputAdapter {
 
     private final RealitiesInputPort realitiesInputPort;
@@ -28,31 +28,21 @@ public class RealityInputAdapter {
     @GetMapping
     public List<RealityResponse> getRealities() {
         List<Reality> realities = realitiesInputPort.getRealities();
-        return realities.parallelStream().map(reality -> new RealityResponse(
-                        reality.getId(),
-                        reality.getType(),
-                        reality.getLocation(),
-                        reality.getPrice(),
-                        reality.getRooms(),
-                        reality.getArea(),
-                        reality.getDescription(),
-                        null,
-                        null
-                ))
-                .collect(Collectors.toList());
+        List<RealityResponse> realityResponseList = RealityInputMapper.mapRealityListToRealityResponseList(realities);
+        return realityResponseList;
     }
 
     @GetMapping("/paginated")
     public Page<RealityResponse> getPage(Pageable page) {
         Page<Reality> realityPage = realitiesInputPort.getRealitiesByPage(page);
         List<Reality> realityList = realityPage.getContent();
-        List<RealityResponse> realityResponseList = RealityMapper.mapRealityListToRealityResponseList(realityList);
+        List<RealityResponse> realityResponseList = RealityInputMapper.mapRealityListToRealityResponseList(realityList);
         return new PageImpl<>(realityResponseList, page, realityResponseList.size());
     }
 
     @GetMapping("/{realityId}")
     public RealityResponse getRealityById(@PathVariable Long realityId) {
-        return RealityMapper.mapRealityToRealityResponse(realitiesInputPort.getRealityById(realityId));
+        return RealityInputMapper.mapRealityToRealityResponse(realitiesInputPort.getRealityById(realityId));
     }
 
     // todo: finish add + update
@@ -62,7 +52,7 @@ public class RealityInputAdapter {
         createRealitiesInputPort.addReality(reality);
     }
 
-    @PostMapping( "/{realityId}")
+    @PostMapping("/{realityId}")
     public void updateReality(@RequestBody Reality reality, @PathVariable Long realityId) throws RealityNotFoundException {
         updateRealitiesInputPort.updateReality(reality, realityId);
     }
